# backend/greenride-api-clean/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GreenRide API Documentation\n## Response Codes:\n- 0000: Success - 操作成功\n- 9999: System Error - 系统错误\n- 1000: Parameter Error - 参数错误\n- 1001: Authorization Error - 授权失败\n- 1002: Authentication Expired - 认证过期\n- 2001: User Error - 用户相关错误\n- 4001: Business Error - 业务错误\n- 4002: Data Expired - 数据过期",
        "title": "GreenRide API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/account/delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "软删除用户账户，添加删除标记和时间戳，保留数据但标记为已删除状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "删除账户",
                "parameters": [
                    {
                        "description": "删除请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户修改密码，需要提供旧密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "用户登录接口，支持邮箱和手机号登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或认证失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户登出，清除服务端相关信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/nearby": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "司机或服务提供者根据当前地理位置获取附近的待接订单列表，支持按距离、订单类型等条件进行筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单管理"
                ],
                "summary": "获取附近的待接订单",
                "parameters": [
                    {
                        "description": "获取附近订单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.GetNearbyOrdersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功，返回附近订单列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/protocol.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/protocol.Order"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "401": {
                        "description": "用户未认证",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "403": {
                        "description": "权限不足，仅司机或服务提供者可访问",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/offline": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "司机下线，只有司机类型用户可以调用此接口。下线后司机状态变为离线，车辆状态变为不可用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "司机"
                ],
                "summary": "司机下线",
                "responses": {
                    "200": {
                        "description": "下线成功",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/online": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "司机上线，只有司机类型用户可以调用此接口。上线后司机状态变为在线，车辆状态变为可用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "司机"
                ],
                "summary": "司机上线",
                "responses": {
                    "200": {
                        "description": "上线成功",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/order/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "服务提供者接受指定的订单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单管理"
                ],
                "summary": "接受订单",
                "parameters": [
                    {
                        "description": "接受订单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.OrderAcceptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "接受失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/order/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取消指定的订单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单管理"
                ],
                "summary": "取消订单",
                "parameters": [
                    {
                        "description": "取消请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.OrderCancelAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取消失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/order/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的订单，支持网约车等类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单管理"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "创建订单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/order/detail": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定订单的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单管理"
                ],
                "summary": "获取订单详情",
                "parameters": [
                    {
                        "description": "订单ID请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.OrderIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/order/estimate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "使用价格引擎预估订单费用，支持价格快照和规则应用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "价格快照"
                ],
                "summary": "预估订单费用V2",
                "parameters": [
                    {
                        "description": "预估请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.EstimateOrderV2Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "预估失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/api/order/rating": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "对订单进行评价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单评价"
                ],
                "summary": "创建评价",
                "parameters": [
                    {
                        "description": "评价请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.CreateOrderRatingAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评价失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/order/ratings": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取订单的评价列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单评价"
                ],
                "summary": "获取评价列表",
                "parameters": [
                    {
                        "description": "订单ID请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.OrderIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/orders": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户的订单列表，支持分页和类型过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单管理"
                ],
                "summary": "获取订单列表",
                "parameters": [
                    {
                        "description": "获取订单列表请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.GetOrdersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户信息",
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/profile/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按需更新用户个人信息，只更新提供的字段",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新个人信息",
                "parameters": [
                    {
                        "description": "更新请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/rating/delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除订单评价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单评价"
                ],
                "summary": "删除评价",
                "parameters": [
                    {
                        "description": "评价ID请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.RatingIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/rating/reply": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "服务提供者回复用户评价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单评价"
                ],
                "summary": "回复评价",
                "parameters": [
                    {
                        "description": "回复请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.ReplyToRatingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "回复失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/rating/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新订单评价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单评价"
                ],
                "summary": "更新评价",
                "parameters": [
                    {
                        "description": "评价更新请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.UpdateOrderRatingAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "description": "用户注册接口，支持乘客和司机注册。country_code为可选参数，默认为卢旺达(RW)。如果提供verify_code，会自动按优先级验证：先验证手机验证码，如果失败再验证邮箱验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或用户已存在",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码",
                    "type": "string"
                },
                "old_password": {
                    "description": "旧密码",
                    "type": "string"
                }
            }
        },
        "handlers.DeleteAccountRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "description": "当前密码，用于确认",
                    "type": "string"
                },
                "reason": {
                    "description": "删除原因（可选）",
                    "type": "string"
                }
            }
        },
...
        "protocol.OrderAcceptRequest": {
            "type": "object",
            "required": [
                "order_id"
            ],
            "properties": {
                "order_id": {
                    "description": "订单ID",
                    "type": "string"
                }
            }
        },
...
        "protocol.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}